refactor(lookup+profile): dedupe helpers, params, auth gate; split profile
Bundles four automated code-quality issues into one focused refactor:

#252 — Promote `_primarySubtag` to public `primaryLanguageSubtag` in
app_language_catalog.dart and delete the duplicate `_primarySubtagOnly`
from lookup_target_languages.dart (6 call sites switched over). The two
helpers were functionally identical; now there is one source of truth.

#254 — Introduce a `LookupTextParams` base class holding the shared
(text, sourceLanguage, targetLanguage) shape + equality, and make the
three concrete params classes extend it. Rename
`LookupDictionaryParams.word` -> `text` so the field matches the other
two classes (only external reader was lookup_section_providers.dart).
Regenerates the provider family hash in lookup_section_providers.g.dart.

#253 — Extract `LookupSectionAuthGate`, a ConsumerWidget that wraps the
repeated `authCtrlProvider.when(...)` + signed-in check +
AuthRequiredCallout fallback used by all three lookup sections. Each
section now wraps its signed-in body in the gate. Also fixes a real UX
drift the issue flagged: DictionaryLookupSection now handles
CreditsFailure with the same "retry + View plans" CTA that
TranslationLookupSection already had (dictionary goes through guardAiCall
which maps HTTP 402 -> CreditsFailure, so it could reach this branch).

#258 — Split profile_content.dart (1064 -> 418 lines) by extracting the
file-private helper widgets into five sibling files:
profile_section_header, profile_hero_card (+ SubscriptionChip),
profile_account_card (+ ProfileNavTile), profile_stats (practice section
+ skeleton + row + tile), profile_sign_out_button. The inline
preferences form stays in profile_content.dart to avoid a state-ownership
behavior change (documented in the PR). ProfileContent public API is
unchanged.

Verification: validate_ci_gates.sh --fix --analyze --test passes (the 2
remaining info-level lints are pre-existing on main, unrelated to this
PR). 124 tests pass across lookup/auth/settings.

Resolves #252
Resolves #253
Resolves #254
Resolves #258
